fixed cleanDB bug-seeding trips

+++ server/seeds/cleanDB.js
@@ -4,11 +4,11 @@ const db = require("../config/connection");
 
 module.exports = async (modelName, collectionName) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    let modelExists = await models[modelName]?.db?.db?.listCollections({
       name: collectionName
     }).toArray()
 
-    if (modelExists.length) {
+    if (modelExists && modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
